GA/NT/Installer: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@63091 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/Installer/VBoxDrvInst.cpp
@@ -1262,8 +1262,8 @@ int __cdecl _tmain(int argc, _TCHAR *argv[])
                 }
                 if (pbVal == NULL) /* By default interpret value as string */
                 {
-                    pbVal = (BYTE*)argv[7];
-                    dwValSize = _tcslen(argv[7]);
+                    pbVal = (BYTE *)argv[7];
+                    dwValSize = (DWORD)_tcslen(argv[7]);
                 }
                 if (argc > 9)
                     dwValSize = _ttol(argv[9]);      /* Get the size in bytes of the value we want to write */
